requirements/base.txt
amqp==5.2.0
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
billiard==4.2.0
celery==5.3.4
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.7
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.3
Django==3.2.9
django-allauth==0.46.0
django-celery-beat==2.2.1
django-cors-headers==3.10.0
django-environ==0.8.1
django-extensions==3.1.5
django-import-export==2.6.1
django-jazzmin==2.6.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.4.0
identify==2.4.0
idna==3.3
kombu==5.4.0
MarkupPy==1.14
nodeenv==1.6.0
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
platformdirs==2.4.0
pre-commit==2.15.0
prompt-toolkit==3.0.23
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0.2
requests==2.26.0
requests-oauthlib==1.3.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
toml==0.10.2
tzdata==2024.1
urllib3==1.26.7
vine==5.1.0
virtualenv==20.10.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
